pyproject.toml
[project]
name = "youtube-whisper-deepseek"
version = "0.1.0"
description = "Download YouTube audio, transcribe with Whisper, and translate subtitles with DeepSeek."
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "faster-whisper>=1.1.1",
    "openai>=1.55.0",
    "python-dotenv>=1.0.1",
    "yt-dlp>=2025.1.15",
]

[project.scripts]
yt-translate = "youtube_translator.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["youtube_translator*"]

[tool.uv]
package = true
